Alors j'ai changé le nom de import en imp_exp. **** J'ai modifié les en-têtes de la fonction import et ça fonctionne. **** J'ai écrit une fonction d'exportation aussi qui fonctionne. j'ai pas mis bcq de commentaires dans les fonctions.
L'écriture est pas trop légère mais je sais pas faire mieux.

+++ code/cercle.c
@@ -1,12 +1,12 @@
-#include <cercle.h>
+#include "cercle.h"
 
-int coord_centre(img I, pixel* p, uint8 c)
+int coord_centre(img I, pixel* p, int c)
 {
     return 0;
 }
 
 
-int compar_cercle(img I, img J, uint8 c)
+int compar_cercle(img I, img J, int c)
 {
     return 0;
 }
