+++ code/cercle.h
@@ -1,13 +1,15 @@
 #ifndef CERCLE
 #define CERCLE
+
+
 #include <stdio.h>
 #include <stdlib.h>
 #include "type.h"
 
-int coord_centre(img I, pixel* p, uint8 c);
+int coord_centre(img I, pixel* p, int c);
 
 
-int compar_cercle(img I, img J, uint8 c);
+int compar_cercle(img I, img J, int c);
 
 
 #endif
